+++ ado_wrapper/dumps.py
@@ -1,5 +1,6 @@
 """This file hosts a collection of dumps with the purpose of documenting the responses, all personal information has been removed."""
 
+# flake8: noqa
 # fmt: off
 
 ADO_USER_DUMP = {
@@ -22,7 +23,7 @@ AUDIT_LOG_DUMP = {
     'correlationId': '00000000-0000-0000-0000-000000000000',
     'activityId': '00000000-0000-0000-0000-000000000000',
     'actorCUID': '00000000-0000-0000-0000-000000000000',
-    'actorUserId': '00000000-0000-0000-0000-000000000000', # Same as above
+    'actorUserId': '00000000-0000-0000-0000-000000000000',  # Same as above
     'actorClientId': '00000000-0000-0000-0000-000000000000',
     'actorUPN': 'first.last@example.com',
     'authenticationMechanism': 'PAT_Unscoped authorizationId:<32_char_uuid>',
@@ -34,7 +35,8 @@ AUDIT_LOG_DUMP = {
     'projectName': None,
     'ipAddress': '128.128.128.128',
     'userAgent': 'VSServices/128.128.123456.0 (NetStandard; Linux 5.10.215-203.850.amzn2.x86_64 #1 SMP Tue Apr 23 20:32:19 UTC 2024) VstsAgentCore-l',
-    'actionId': 'Library.AgentAdded', 'data': {...},
+    'actionId': 'Library.AgentAdded',
+    'data': {...},
     'details': 'Added agent <agent_name> to pool <pool_name>.',
     'area': 'Library',
     'category': 'modify',
@@ -93,9 +95,9 @@ BUILD_DEFINITION_DUMP = {
     "description": "<description>",
     "jobAuthorizationScope": "projectCollection",
     "process": {"yamlFilename": "build.yaml", "type": 2},
-    "repository": {"id": "<repo_id>","type": "TfsGit", "name": "ado_wrapper-test-repo-for-create-delete-builds", "url": "https://dev.azure.com/{ado_client.org}/{ado_client.project}/_git/ado_wrapper-test-repo-for-create-delete-builds", "defaultBranch": "refs/heads/my-branch", "clean": None, "checkoutSubmodules": False},
+    "repository": {"id": "<repo_id>", "type": "TfsGit", "name": "ado_wrapper-test-repo-for-create-delete-builds", "url": "https://dev.azure.com/{ado_client.org}/{ado_client.project}/_git/ado_wrapper-test-repo-for-create-delete-builds", "defaultBranch": "refs/heads/my-branch", "clean": None, "checkoutSubmodules": False},
     "quality": "definition",
-    "authoredBy": {"displayName": "<First> <Last>", "url": "https://spsprodweu5.vssps.visualstudio.com/<some_id>/_apis/Identities/<user_id>", "_links": {"avatar": {"href": "https://dev.azure.com/{ado_client.org}/_apis/GraphProfile/MemberAvatars/<user_descriptor>"}},"id": "<user_id>", "uniqueName": "<first.last@company.com>", "imageUrl": "https://dev.azure.com/{ado_client.org}/_apis/GraphProfile/MemberAvatars/<user_descriptor>", "descriptor": "<user_descriptor>"},
+    "authoredBy": {"displayName": "<First> <Last>", "url": "https://spsprodweu5.vssps.visualstudio.com/<some_id>/_apis/Identities/<user_id>", "_links": {"avatar": {"href": "https://dev.azure.com/{ado_client.org}/_apis/GraphProfile/MemberAvatars/<user_descriptor>"}}, "id": "<user_id>", "uniqueName": "<first.last@company.com>", "imageUrl": "https://dev.azure.com/{ado_client.org}/_apis/GraphProfile/MemberAvatars/<user_descriptor>", "descriptor": "<user_descriptor>"},
     "drafts": [],
     "queue": {"_links": {"self": {"href": "https://dev.azure.com/{ado_client.org}/_apis/build/Queues/575"}}, "id": 575, "name": "<pool_name>", "url": "https://dev.azure.com/{ado_client.org}/_apis/build/Queues/575", "pool": {"id": 195, "name": "<pool_name>"}},
     "id": 2601,
@@ -168,12 +170,13 @@ PULL_REQUEST_DUMP = {
     "artifactId": "vstfs:///Git/PullRequestId/<project_id>%2f<repo_id>%2f10237"
 }
 
-PULL_REQUEST_SET_ASSIGNED_TO_MY_TEAM = {"pullRequestListCustomCriteria":
-    [
-        {"authorIds": ["<author_id>"], "groupByVote": False, "id": "CreatedByMe", "includeDuplicates": False, "readonly": True, "status": 1, "title":"Created by me"},
-        {"groupByVote": True, "id": "AssignedToMe", "includeDuplicates": False, "readonly": True, "reviewerIds": ["<author_id>"], "status": 1, "title": "Assigned to me"},
-        {"authorIds": [], "id": "AssignedToMyTeams", "myTeamsAsReviewer": True, "status": 2, "title": "Assigned to my teams"}
-    ]
+PULL_REQUEST_SET_ASSIGNED_TO_MY_TEAM = {
+    "pullRequestListCustomCriteria":
+        [
+            {"authorIds": ["<author_id>"], "groupByVote": False, "id": "CreatedByMe", "includeDuplicates": False, "readonly": True, "status": 1, "title": "Created by me"},
+            {"groupByVote": True, "id": "AssignedToMe", "includeDuplicates": False, "readonly": True, "reviewerIds": ["<author_id>"], "status": 1, "title": "Assigned to me"},
+            {"authorIds": [], "id": "AssignedToMyTeams", "myTeamsAsReviewer": True, "status": 2, "title": "Assigned to my teams"}
+        ]
 }
 
 REPO_DUMP = {
@@ -196,7 +199,7 @@ REPO_POLICY_DUMP = {
             'identities': "<Member>", 'supportServicePrincipals': True, 'isEditable': True, 'buildDefinitions': None, 'recentStatuses': None,
             'policyGroups': {'{policy_id_uuid}': {
                              'currentScopePolicies': [{'createdBy': "<Member>", 'createdDate': '/Date(1712832031819)/',
-                                                       'isEnabled': True, 'isBlocking': True, 'isDeleted': False, 'settings': {'minimumApproverCount': 1, 'creatorVoteCounts': False, 'allowDownvotes': False, 'resetOnSourcePush': False, 'requireVoteOnLastIteration': False, 'resetRejectionsOnSourcePush': False, 'blockLastPusherVote': True, 'requireVoteOnEachIteration': False, 'scope': [{'refName': 'refs/heads/main', 'matchKind': 'Exact', 'repositoryId': '{repo_id}'}]}, 'isEnterpriseManaged': False, '_links': "<links>", 'revision': 13, 'id': 8178, 'url': 'https://dev.azure.com/{ado_org}/{repo_id}/_apis/policy/configurations/8178', 'type': {'id': '{policy_id_uuid}', 'url': 'https://dev.azure.com/{ado_org}/{repo_id}/_apis/policy/types/{policy_id_uuid}', 'displayName': 'Minimum number of reviewers'}}], 'enterpriseManagedPolicies': None, 
+                                                       'isEnabled': True, 'isBlocking': True, 'isDeleted': False, 'settings': {'minimumApproverCount': 1, 'creatorVoteCounts': False, 'allowDownvotes': False, 'resetOnSourcePush': False, 'requireVoteOnLastIteration': False, 'resetRejectionsOnSourcePush': False, 'blockLastPusherVote': True, 'requireVoteOnEachIteration': False, 'scope': [{'refName': 'refs/heads/main', 'matchKind': 'Exact', 'repositoryId': '{repo_id}'}]}, 'isEnterpriseManaged': False, '_links': "<links>", 'revision': 13, 'id': 8178, 'url': 'https://dev.azure.com/{ado_org}/{repo_id}/_apis/policy/configurations/8178', 'type': {'id': '{policy_id_uuid}', 'url': 'https://dev.azure.com/{ado_org}/{repo_id}/_apis/policy/types/{policy_id_uuid}', 'displayName': 'Minimum number of reviewers'}}], 'enterpriseManagedPolicies': None,
                                                        'inheritedPolicies': [{'createdBy': "<Member>", 'createdDate': '/Date(1710074353140)/', 'isEnabled': True, 'isBlocking': True, 'isDeleted': False, 'settings': {'minimumApproverCount': 1, 'creatorVoteCounts': False, 'allowDownvotes': False, 'resetOnSourcePush': False, 'requireVoteOnLastIteration': False, 'resetRejectionsOnSourcePush': False, 'blockLastPusherVote': False, 'requireVoteOnEachIteration': False, 'scope': [{'refName': None, 'matchKind': 'DefaultBranch', 'repositoryId': None}]}, 'isEnterpriseManaged': False, '_links': {'self': {'href': 'https://dev.azure.com/{ado_org}/{repo_id}/_apis/policy/configurations/{config_id}'}, 'policyType': {'href': 'https://dev.azure.com/{ado_org}/{repo_id}/_apis/policy/types/{policy_id_uuid}'}}, 'revision': 2, 'id': "{config_id}", 'url': 'https://dev.azure.com/{ado_org}/{repo_id}/_apis/policy/configurations/{config_id}', 'type': {'id': '{policy_id_uuid}', 'url': 'https://dev.azure.com/{ado_org}/{repo_id}/_apis/policy/types/{policy_id_uuid}', 'displayName': 'Minimum number of reviewers'}}]}, 'c6a1889d-b943-4856-b76f-9e46bb6b0df2': {'currentScopePolicies': None, 'enterpriseManagedPolicies': None,
                                                         'inheritedPolicies': [{'createdBy': "<Member>", 'createdDate': '/Date(1670239096642)/', 'isEnabled': True, 'isBlocking': True, 'isDeleted': False, 'settings': {'scope': [{'refName': None, 'matchKind': 'DefaultBranch', 'repositoryId': None}]}, 'isEnterpriseManaged': False, '_links': {'self': {'href': 'https://dev.azure.com/{ado_org}/{repo_id}/_apis/policy/configurations/40'}, 'policyType': {'href': 'https://dev.azure.com/{ado_org}/{repo_id}/_apis/policy/types/c6a1889d-b943-4856-b76f-9e46bb6b0df2'}}, 'revision': 1, 'id': 40, 'url': 'https://dev.azure.com/{ado_org}/{repo_id}/_apis/policy/configurations/40', 'type': {'id': 'c6a1889d-b943-4856-b76f-9e46bb6b0df2', 'url': 'https://dev.azure.com/{ado_org}/{repo_id}/_apis/policy/types/c6a1889d-b943-4856-b76f-9e46bb6b0df2', 'displayName': 'Comment requirements'}}]},
                                                         'fd2167ab-b0be-447a-8ec8-39368250530e': {'currentScopePolicies': [{'createdBy': "<Member>", 'createdDate': '/Date(1712766957032)/', 'isEnabled': True, 'isBlocking': True, 'isDeleted': False, 'settings': {'requiredReviewerIds': ['ab35e0e5-b36d-46c4-8d91-714d413e4fae'], 'minimumApproverCount': 1, 'creatorVoteCounts': True, 'scope': [{'refName': 'refs/heads/main', 'matchKind': 'Exact', 'repositoryId': '1aa43e9c-ffca-4388-98b5-59b272a497b8'}]}, 'isEnterpriseManaged': False, '_links': {'self': {'href': 'https://dev.azure.com/{ado_org}/{repo_id}/_apis/policy/configurations/8179'}, 'policyType': {'href': 'https://dev.azure.com/{ado_org}/{repo_id}/_apis/policy/types/fd2167ab-b0be-447a-8ec8-39368250530e'}}, 'revision': 1, 'id': 8179, 'url': 'https://dev.azure.com/{ado_org}/{repo_id}/_apis/policy/configurations/8179', 'type': {'id': 'fd2167ab-b0be-447a-8ec8-39368250530e', 'url': 'https://dev.azure.com/{ado_org}/{repo_id}/_apis/policy/types/fd2167ab-b0be-447a-8ec8-39368250530e', 'displayName': 'Required reviewers'}}],
@@ -214,6 +217,26 @@ RUN_DUMP = {
     'resources': {'repositories': {'self': {'repository': {'id': '<repo_id>', 'type': 'azureReposGit'}, 'refName': 'refs/heads/my-branch', 'version': 'bf9b3ae18fa2bd5c0f6c99f7ad67f5e58e0a6a8e'}}},
     'id': "<run_id>",
     'name': '20240609.1',
+}
+
+RUN_RESULT_DUMP = {
+    "id": "a8c942cb-b48a-5d20-399e-e9c5aed77b37",
+    "name": "Build",
+    "imageName": "Ubuntu-20.04",
+    "queueId": 36,
+    "stageName": "stage name with spaces",
+    "stageId": "<32char_uuid>",
+    "startTime": "/Date(1720168453910)/",
+    "finishTime": "/Date(1720168463150)/",
+    "stateData": {
+        "pendingDependencies": False,
+        "pendingChecks": False,
+    },
+    "linkProps": {
+        "href": "/{ado_org}/{ado_project_id}}/_build/results?buildId={build_id}&view=logs&j=a8c942cb-b48a-5d20-399e-e9c5aed77b37"
+    },
+    "state": 2,
+    "result": 0
 }
 
 SERVICE_ENDPOINT_DUMP = {