+++ ado_wrapper/generate_docs.py
@@ -1,7 +1,7 @@
 import inspect
 import re
 
-from ado_wrapper.resources import *  # pylint: disable=W0401,W0614
+from ado_wrapper.resources import *  # pylint: disable=W0401,W0614  # noqa: F401,F403
 
 pattern = re.compile(r"(?<!^)(?=[A-Z])")
 ignored_functions = ["get_by_url", "to_json", "from_json", "get_by_abstract_filter", "from_request_payload", "set_lifecycle_policy"]